# PoliTechAPI/openapi.generated.yaml
openapi: 3.0.3
info:
  title: PoliTech API (generated from controllers)
  version: 1.0.0
  description: >
    OpenAPI specification built from Spring controllers.
    WARNING: Some endpoints return placeholder/null responses in code.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /api/v1/{tenantCode}/dashboard/chart:
    get:
      tags: [dashboard]
      summary: Get dashboard chart
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
        - name: period
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Chart response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardChartResponse"
      security:
        - bearerAuth: []
  /api/v1/{tenantCode}/dashboard/cards:
    get:
      tags: [dashboard]
      summary: Get dashboard cards
      parameters:
        - name: tenantCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cards response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardCardsResponse"
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DashboardChartResponse:
      type: object
      description: Chart response
      x-java-class: ru.pt.api.dto.dashboard.DashboardChartResponse
    DashboardCardsResponse:
      type: object
      description: Cards response
      x-java-class: ru.pt.api.dto.dashboard.DashboardCardsResponse
...